introduction of new ItemCapability to:  - Optimise thermometer property rendering and fix the actual crash  - Optimise the world temperature calcul when right clicking the thermometer  - Optimise the world temperature calcul when rendering in item frame Improvement on the entity looked at calculation, reduce refresh rate by 5 Visual indication when player target temperature isn't the actual player temperature Handle the temporary effects of food using the Minecraft effect system + config update Fix + improvement of the shade and whether temperature influence Improve the fern's spawn Refactor of timing usage (system current millis) to use proper game ticks (client, player + world age) Refactor the distance influence in the block modifier calculation

+++ src/main/java/sfiomn/legendarysurvivaloverhaul/api/temperature/TemperatureUtil.java
@@ -16,7 +16,7 @@ public class TemperatureUtil
 	 * @param player Player
 	 * @return Player's target temperature
 	 */
-	public static int getPlayerTargetTemperature(PlayerEntity player)
+	public static float getPlayerTargetTemperature(PlayerEntity player)
 	{
 		return internal.getPlayerTargetTemperature(player);
 	}
@@ -28,7 +28,7 @@ public class TemperatureUtil
 	 * @param pos Position
 	 * @return World temperature at position
 	 */
-	public static int getWorldTemperature(World world, BlockPos pos)
+	public static float getWorldTemperature(World world, BlockPos pos)
 	{
 		return internal.getWorldTemperature(world, pos);
 	}
@@ -38,7 +38,7 @@ public class TemperatureUtil
 	 * @param temperature Temperature
 	 * @return Clamped temperature
 	 */
-	public static int clampTemperature(int temperature)
+	public static float clampTemperature(float temperature)
 	{
 		return internal.clampTemperature(temperature);
 	}
@@ -48,7 +48,7 @@ public class TemperatureUtil
 	 * @param temperature Temperature
 	 * @return TemperatureEnum for the temperature
 	 */
-	public static TemperatureEnum getTemperatureEnum(int temperature)
+	public static TemperatureEnum getTemperatureEnum(float temperature)
 	{
 		return internal.getTemperatureEnum(temperature);
 	}
